Debugged counting errors and implemented a cutoff index.
Some mutations we no longer wish to count, such as those
occuring in regions of type 'None' (no annotations) or those
occuring after a specified cutoff. Both have been dealt with in
this commit. Cutoff is currently hardcoded and calculated by
subtracting 400 from the end of the template CDS index.